upstreaming subpixel_refinement_utils
